requirements.txt
python
python-dotenv
Node.js
npm
sqlalchemy
fastapi
azure-storage-blob
azure-identity
pandas
uvicorn[standard]